+++ src/components/Prescription/PrescriptionPreview.tsx
@@ -14,13 +14,8 @@ import { formatDosage, formatSig } from "@/components/Medicine/utils";
 import { formatDateTime, formatName, formatPatientAge } from "@/Utils/utils";
 import useCurrentFacility from "@/pages/Facility/utils/useCurrentFacility";
 import { displayMedicationName } from "@/types/emr/medicationRequest/medicationRequest";
-import { PatientRead } from "@/types/emr/patient/patient";
 import { PrescriptionRead } from "@/types/emr/prescription/prescription";
-import { FacilityBareMinimum, FacilityBase } from "@/types/facility/facility";
-import {
-  PatientIdentifier,
-  PatientIdentifierUse,
-} from "@/types/patient/patientIdentifierConfig/patientIdentifierConfig";
+import { PatientIdentifierUse } from "@/types/patient/patientIdentifierConfig/patientIdentifierConfig";
 import { QRCodeSVG } from "qrcode.react";
 
 interface DetailRowProps {
@@ -35,167 +30,6 @@ const PrescriptionContent = ({
   prescription: PrescriptionRead;
 }) => {
   const medications = prescription.medications;
-  const { t } = useTranslation();
-
-  return (
-    <div>
-      {/* Prescription Symbol */}
-      <div className="text-xl font-semibold mb-3 flex items-end gap-4">
-        <p>{t("℞")}</p>
-        <p className="text-sm text-gray-600 font-semibold ">
-          {formatDateTime(prescription.created_date, "DD/MM/YYYY hh:mm A")}
-        </p>
-      </div>
-
-      {/* Medications Table */}
-      {medications && medications.length > 0 && (
-        <div className="mt-4">
-          <p className="text-base font-semibold mb-2">{t("medicines")}</p>
-          <PrintTable
-            headers={[
-              { key: "medicine" },
-              { key: "dosage" },
-              { key: "frequency" },
-              { key: "duration" },
-              { key: "instructions" },
-            ]}
-            rows={medications?.map((medication) => {
-              const instruction = medication.dosage_instruction[0];
-              const frequency = getFrequencyDisplay(instruction?.timing);
-              const dosage = formatDosage(instruction);
-              const duration = instruction?.timing?.repeat?.bounds_duration;
-              const remarks = formatSig(instruction);
-              const notes = medication.note;
-              return {
-                medicine: displayMedicationName(medication),
-                status: t(`medication_status_${medication.status}`),
-                dosage: dosage,
-                frequency: instruction?.as_needed_boolean
-                  ? `${t("as_needed_prn")}`
-                  : (frequency?.meaning ?? "-") +
-                    (instruction?.additional_instruction?.[0]?.display
-                      ? `, ${instruction.additional_instruction[0].display}`
-                      : ""),
-                duration: duration ? `${duration.value} ${duration.unit}` : "-",
-                instructions: [remarks, notes].filter(Boolean).join("\n"),
-              };
-            })}
-            className="text-xs font-semibold whitespace-break-spaces text-gray-950"
-            cellConfig={{
-              medicine: { className: "text-left" },
-            }}
-          />
-        </div>
-      )}
-      {prescription?.note && (
-        <div className="mt-6 mb-6 text-sm text-gray-600">
-          <p className="font-semibold mb-1">{t("note")}</p>
-          <Markdown
-            content={prescription.note}
-            prose={false}
-            className="text-sm"
-          />
-        </div>
-      )}
-      {/* Doctor's Signature */}
-      <div className="w-full items-end mt-6 flex flex-row justify-end gap-1">
-        <div className="text-right">
-          <p className="text-xs text-gray-400">{t("prescribed_by")}</p>
-          <p className="text-sm text-gray-600 font-semibold">
-            {formatName(prescription.prescribed_by)}
-          </p>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-export const PrescriptionPrintHeader = ({
-  patient,
-  facility,
-}: {
-  patient: PatientRead;
-  facility?: (FacilityBareMinimum & Partial<FacilityBase>) | null;
-}) => {
-  const { t } = useTranslation();
-
-  return (
-    <>
-      <div className="mb-4 flex justify-between items-start border-b border-gray-200 pb-2">
-        <div className="flex items-start gap-4">
-          <div className="text-left">
-            <h1 className="text-2xl font-medium">{facility?.name || ""}</h1>
-            {facility?.address && (
-              <div className="text-sm rounded-md whitespace-pre-wrap text-gray-500">
-                {facility.address}
-                {facility.phone_number && (
-                  <p className="text-sm text-gray-500">
-                    {t("phone")}: {facility.phone_number}
-                  </p>
-                )}
-              </div>
-            )}
-          </div>
-        </div>
-        <QRCodeSVG value={patient.id} size={50} level="Q" marginSize={0} />
-        <img
-          src={careConfig.mainLogo?.dark}
-          alt="Logo"
-          className="mb-2 h-10 w-auto object-contain sm:mb-0 text-end"
-        />
-      </div>
-
-      <div className="mb-6 grid grid-cols-1 gap-6 md:grid-cols-2 print:grid-cols-2">
-        <div className="space-y-1">
-          <DetailRow label={t("patient")} value={patient.name} isStrong />
-          <DetailRow
-            label={`${t("age")} / ${t("sex")}`}
-            value={
-              patient
-                ? `${formatPatientAge(patient, true)}, ${t(
-                    `GENDER__${patient.gender}`,
-                  )}`
-                : undefined
-            }
-            isStrong
-          />
-          {patient.instance_identifiers
-            ?.filter(
-              ({ config }: PatientIdentifier) =>
-                config.config.use === PatientIdentifierUse.official,
-            )
-            .map((identifier: PatientIdentifier) => (
-              <DetailRow
-                key={identifier.config.id}
-                label={identifier.config.config.display}
-                value={identifier.value}
-                isStrong
-              />
-            ))}
-        </div>
-        <div className="space-y-1">
-          <DetailRow
-            label={t("date")}
-            value={format(new Date(), "dd MMM yyyy, EEEE")}
-            isStrong
-          />
-          <DetailRow
-            label={t("mobile_number")}
-            value={patient && formatPhoneNumberIntl(patient.phone_number)}
-            isStrong
-          />
-        </div>
-      </div>
-    </>
-  );
-};
-
-export const PrescriptionDetails = ({
-  prescription,
-}: {
-  prescription: PrescriptionRead;
-}) => {
-  const medications = prescription.medications || [];
   const { t } = useTranslation();
 
   return (
